dom/indexedDB/test/test_filehandle_getFile.html

<!--
  Any copyright is dedicated to the Public Domain.
  http://creativecommons.org/publicdomain/zero/1.0/
-->
<html>
<head>
  <title>File Handle Test</title>

  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />

  <script type="text/javascript;version=1.7">
  function testSteps()
  {
    const name = window.location.pathname;

    let request = indexedDB.open(name, 1);
    request.onerror = errorHandler;
    request.onsuccess = grabEventAndContinueHandler;
    let event = yield undefined;

    let db = event.target.result;
    db.onerror = errorHandler;

    request = db.createMutableFile("test.txt");
    request.onerror = errorHandler;
    request.onsuccess = grabEventAndContinueHandler;
    event = yield undefined;

    let mutableFile = event.target.result;
    mutableFile.onerror = errorHandler;

    request = mutableFile.getFile();
    ok(request instanceof DOMRequest, "Correct interface");
    ok(!(request instanceof IDBFileRequest), "Correct interface");
    ok(!('fileHandle' in request), "Property should not exist");
    ok(request.fileHandle === undefined, "Property should not exist");
    ok(!('lockedFile' in request), "Property should not exist");
    ok(request.lockedFile === undefined, "Property should not exist");
    ok(!('onprogress' in request), "Property should not exist");
    ok(request.onprogress === undefined, "Property should not exist");

    finishTest();
    yield undefined;
  }
  </script>
  <script type="text/javascript;version=1.7" src="file.js"></script> <!-- This is included Just to skip this test in test_ipc.html -->
  <script type="text/javascript;version=1.7" src="helpers.js"></script>

</head>

<body onload="runTest();"></body>

</html>
